+++ ted2zim/multi/scraper.py
@@ -53,8 +53,10 @@ class TedHandler(object):
         return [sys.argv[0].replace(f"{cmd}-multi", cmd)]
 
     @staticmethod
-    def compute_format(item, fmt):
-        return fmt.format(identity=item.replace(" ", "-"), period="{}")
+    def compute_format(item, slug, fmt):
+        return fmt.format(
+            identity=item.replace(" ", "-"), period="{}", slug=slug.replace("_", "-")
+        )
 
     @staticmethod
     def download_playlists_list_from_site(topics_list):
@@ -224,11 +226,15 @@ class TedHandler(object):
                 "--topics",
                 item,
             ]
+            slug = item.replace(" ", "-")
         elif mode == "playlist":
             args += [
                 "--playlist",
                 item,
             ]
+            slug = requests.get(f"https://www.ted.com/playlists/{item}").url.replace(
+                f"https://www.ted.com/playlists/{item}/", ""
+            )
         else:
             raise ValueError(f"Unsupported mode {mode}")
 
@@ -248,7 +254,7 @@ class TedHandler(object):
             )
 
             if value:  # only set arg if we have a value so it can be defaulted
-                args += [f"--{key}", self.compute_format(item, str(value))]
+                args += [f"--{key}", self.compute_format(item, slug, str(value))]
 
         # append regular ted2zim args
         args += self.extra_args